Reverted changes when tmpl.go is genereated using go-bindata
