--- ros-gen-go/msg.partial.tmpl.go
@@ -1,92 +1,0 @@
-package main
-
-import (
-	"os"
-	"time"
-)
-
-func msgPartialTmplBytes() ([]byte, error) {
-	return bindataRead(
-		_msgPartialTmpl,
-		"msg.partial.tmpl",
-	)
-}
-
-func msgPartialTmpl() (*asset, error) {
-	bytes, err := msgPartialTmplBytes()
-	if err != nil {
-		return nil, err
-	}
-
-	info := bindataFileInfo{name: "msg.partial.tmpl", size: int64(len(bytes)), mode: os.FileMode(436), modTime: time.Unix(1507566040, 0)}
-	a := &asset{bytes: bytes, info: info}
-	return a, nil
-}
-
-var _msgPartialTmpl = []byte{
-  0x1f, 0x8b, 0x08, 0x08, 0xbf, 0xbd, 0xe4, 0x59, 0x00, 0x03, 0x6d, 0x73,
-  0x67, 0x2e, 0x70, 0x61, 0x72, 0x74, 0x69, 0x61, 0x6c, 0x2e, 0x74, 0x6d,
-  0x70, 0x6c, 0x00, 0xa5, 0x54, 0x6d, 0x4f, 0xdb, 0x40, 0x0c, 0xfe, 0x9c,
-  0xfc, 0x0a, 0xaf, 0x12, 0x28, 0x29, 0x5d, 0x8a, 0x36, 0xf1, 0xa5, 0xac,
-  0x93, 0x26, 0xc1, 0xa6, 0x49, 0x03, 0x26, 0x40, 0xda, 0x07, 0x84, 0xe0,
-  0xd6, 0x38, 0xe9, 0x89, 0xbc, 0xa0, 0xcb, 0x85, 0x50, 0xa2, 0xfe, 0xf7,
-  0xd9, 0x77, 0x69, 0x68, 0x4a, 0x5b, 0xd8, 0x56, 0x09, 0x11, 0xfb, 0xec,
-  0xc7, 0x8f, 0xcf, 0x8f, 0x4f, 0xcf, 0xee, 0x11, 0x6e, 0x4e, 0x8a, 0xb8,
-  0xae, 0x21, 0x38, 0x15, 0x29, 0xc2, 0x7c, 0x0e, 0x85, 0x56, 0xe5, 0x44,
-  0x43, 0xed, 0x3a, 0x1a, 0x1f, 0x35, 0x9b, 0x32, 0x8b, 0x5d, 0x27, 0xe3,
-  0xe3, 0x85, 0x91, 0x86, 0x07, 0x45, 0x99, 0x2e, 0xcc, 0xb9, 0xeb, 0x46,
-  0x65, 0x36, 0x01, 0x4f, 0x43, 0x7f, 0x05, 0xcd, 0x87, 0x4b, 0x02, 0xf1,
-  0xfc, 0x26, 0x94, 0x51, 0x15, 0xea, 0x52, 0x65, 0xa0, 0x03, 0x86, 0xdf,
-  0x9e, 0xcb, 0x1f, 0xeb, 0x73, 0x99, 0xcd, 0xf6, 0xdc, 0x93, 0xa3, 0x83,
-  0x8b, 0x32, 0x5d, 0x9f, 0x6d, 0xe9, 0xbf, 0x52, 0x1b, 0xab, 0x13, 0x2c,
-  0x0a, 0x11, 0x33, 0x03, 0x95, 0x17, 0x41, 0x63, 0x31, 0x50, 0x0a, 0xa3,
-  0x31, 0x64, 0x58, 0x79, 0xcb, 0x19, 0xae, 0x53, 0xd7, 0xc3, 0x3e, 0x9c,
-  0x9e, 0x5d, 0x1e, 0x8f, 0xe0, 0x34, 0xa7, 0x73, 0x0c, 0x41, 0xe7, 0x20,
-  0x33, 0xa9, 0xa5, 0x48, 0xe4, 0x13, 0x42, 0x24, 0x31, 0x09, 0x0b, 0x10,
-  0x05, 0xe8, 0x29, 0xce, 0x40, 0x28, 0x04, 0x91, 0x24, 0xf0, 0x84, 0x2a,
-  0x87, 0x07, 0x91, 0x94, 0x38, 0x80, 0x6a, 0x2a, 0x27, 0x53, 0x90, 0x05,
-  0x84, 0x18, 0x89, 0x32, 0xd1, 0x94, 0x0e, 0x71, 0x0e, 0xfd, 0xe1, 0x7c,
-  0xde, 0x36, 0x60, 0x98, 0x3f, 0x08, 0x05, 0x9e, 0xeb, 0xac, 0x0c, 0x6f,
-  0x0c, 0xbb, 0xab, 0xf3, 0x24, 0xbe, 0xce, 0x2d, 0x3b, 0xce, 0x45, 0x45,
-  0xf6, 0xed, 0x80, 0xec, 0x1e, 0xdb, 0x3f, 0xc5, 0xe4, 0x8e, 0x1a, 0x6a,
-  0xe2, 0x86, 0x4b, 0x39, 0xbd, 0x36, 0xc6, 0xde, 0x62, 0xe3, 0x9a, 0xbb,
-  0xbe, 0xeb, 0x6a, 0xd6, 0xcc, 0x7a, 0xbd, 0xd4, 0xf5, 0x7b, 0x50, 0x22,
-  0xa3, 0x3b, 0x0a, 0xbe, 0xda, 0x4e, 0x99, 0x74, 0x87, 0x4b, 0x0d, 0x32,
-  0x82, 0xe0, 0x7b, 0xf1, 0x45, 0x29, 0x31, 0x23, 0xcf, 0x95, 0xf5, 0xc4,
-  0x1a, 0x02, 0xe3, 0xba, 0xe0, 0x6b, 0xda, 0xa7, 0x03, 0xce, 0x7a, 0xf6,
-  0x18, 0x1b, 0xb3, 0x90, 0x3e, 0xae, 0xdb, 0x2f, 0x0e, 0xf9, 0x96, 0x5f,
-  0x12, 0x9f, 0x06, 0xde, 0x32, 0xb0, 0x87, 0xcf, 0xc3, 0x4d, 0xa1, 0xdf,
-  0x19, 0xec, 0x05, 0x2a, 0x3b, 0x0e, 0xaf, 0x02, 0x99, 0x07, 0xbf, 0x94,
-  0xd4, 0xa8, 0x7c, 0xf0, 0x50, 0x29, 0xa0, 0xbf, 0x9c, 0xbe, 0x37, 0xf5,
-  0x02, 0xc0, 0xee, 0x6e, 0x07, 0xae, 0x33, 0x1c, 0x82, 0x01, 0x81, 0x82,
-  0xa9, 0x26, 0x52, 0xeb, 0x04, 0x99, 0x85, 0x14, 0x19, 0x65, 0x30, 0xec,
-  0x18, 0x7e, 0xcb, 0x4c, 0xa8, 0x99, 0x2d, 0xe6, 0x55, 0x83, 0x85, 0xfd,
-  0xc3, 0x04, 0x1f, 0x9b, 0xd8, 0x01, 0x94, 0x32, 0xd3, 0x1f, 0x3f, 0x78,
-  0x09, 0x66, 0x5e, 0x1a, 0x2c, 0x73, 0xf6, 0x49, 0x5c, 0x54, 0x95, 0xa1,
-  0xde, 0x91, 0xf0, 0x64, 0x62, 0x86, 0xda, 0xa8, 0x21, 0x4a, 0x75, 0x70,
-  0xcc, 0xbc, 0x23, 0xaf, 0x37, 0xc9, 0xcb, 0x24, 0x84, 0x2c, 0xd7, 0x50,
-  0x19, 0x46, 0xc2, 0x90, 0x24, 0xc0, 0x58, 0x4f, 0x47, 0xb0, 0x53, 0xf4,
-  0x06, 0x0c, 0xe2, 0xf3, 0x28, 0x01, 0xa2, 0x5c, 0xc1, 0x0d, 0x39, 0x12,
-  0x34, 0x7a, 0xb6, 0xdd, 0x76, 0x0a, 0x53, 0x19, 0xa0, 0x5f, 0x53, 0x7a,
-  0x6c, 0xd6, 0xa0, 0xbd, 0xbe, 0x66, 0x1f, 0xcc, 0xed, 0x70, 0x4b, 0x46,
-  0x2f, 0x4b, 0xd3, 0xa0, 0x5a, 0xbb, 0x8c, 0xed, 0x1f, 0x76, 0x89, 0x83,
-  0xf9, 0x35, 0xe4, 0xe9, 0xc4, 0x38, 0xe6, 0x96, 0x92, 0x99, 0x5f, 0x52,
-  0xa0, 0xbd, 0xec, 0x7f, 0xaf, 0xdb, 0xbd, 0xbe, 0x35, 0x04, 0x3a, 0xe5,
-  0xdb, 0xca, 0x56, 0x39, 0x4e, 0xab, 0xb0, 0xc5, 0xc6, 0x6d, 0x16, 0xd3,
-  0x11, 0x16, 0xad, 0x9c, 0x14, 0xcb, 0xe9, 0x1c, 0x45, 0xf8, 0x56, 0x39,
-  0xb1, 0x72, 0x96, 0xd0, 0x6c, 0xd4, 0x8a, 0xba, 0xc0, 0x10, 0xa6, 0x40,
-  0x06, 0x5e, 0xa7, 0x30, 0x87, 0x1f, 0x02, 0xe3, 0xb7, 0xf2, 0x69, 0x95,
-  0xd2, 0x8a, 0x8e, 0x33, 0x3d, 0xb5, 0x41, 0x73, 0xbb, 0x9c, 0xba, 0xe5,
-  0x86, 0x3a, 0xea, 0x12, 0x19, 0x4b, 0x4b, 0x31, 0x13, 0xab, 0x2c, 0x53,
-  0x97, 0x75, 0xb4, 0xd4, 0xc6, 0x8a, 0xce, 0x16, 0x4d, 0xbd, 0x5c, 0x71,
-  0xc3, 0xd4, 0x20, 0x7c, 0x86, 0xd5, 0x75, 0xdf, 0x24, 0xf0, 0xcd, 0x65,
-  0xe9, 0xa9, 0xcd, 0x21, 0x11, 0x2a, 0xc6, 0x11, 0xe0, 0xe3, 0x3d, 0x4e,
-  0x34, 0x86, 0xe3, 0x55, 0xd8, 0x01, 0xbd, 0xa5, 0x7a, 0xbc, 0x13, 0x12,
-  0x3f, 0xd3, 0x77, 0x4b, 0x70, 0x21, 0x3a, 0xa7, 0xab, 0xff, 0x31, 0xa4,
-  0xe2, 0x0e, 0xbd, 0xab, 0xeb, 0x17, 0x6f, 0xcd, 0x80, 0x1e, 0x66, 0xed,
-  0x19, 0x10, 0xbf, 0xf3, 0xee, 0x38, 0xcc, 0x4b, 0xf2, 0x42, 0xed, 0x1f,
-  0xd2, 0xff, 0x4f, 0xcf, 0x71, 0x64, 0xee, 0xed, 0xad, 0x5b, 0xa9, 0x25,
-  0x11, 0x75, 0xc4, 0xad, 0xde, 0x20, 0xee, 0x2b, 0x79, 0xfd, 0x37, 0x0b,
-  0x46, 0x42, 0x76, 0xb6, 0x2c, 0xd9, 0xff, 0x30, 0x79, 0xc3, 0x9a, 0x39,
-  0xaf, 0x6d, 0xd9, 0x1f, 0xb8, 0xb1, 0x60, 0x8c, 0x91, 0x08, 0x00, 0x00 }
-
