--- ros-gen-go/msg.tmpl.go
@@ -1,47 +1,0 @@
-package main
-
-import (
-	"os"
-	"time"
-)
-
-func msgTmplBytes() ([]byte, error) {
-	return bindataRead(
-		_msgTmpl,
-		"msg.tmpl",
-	)
-}
-
-func msgTmpl() (*asset, error) {
-	bytes, err := msgTmplBytes()
-	if err != nil {
-		return nil, err
-	}
-
-	info := bindataFileInfo{name: "msg.tmpl", size: int64(len(bytes)), mode: os.FileMode(436), modTime: time.Unix(1507566040, 0)}
-	a := &asset{bytes: bytes, info: info}
-	return a, nil
-}
-
-var _msgTmpl = []byte{
-  0x1f, 0x8b, 0x08, 0x08, 0x3b, 0xe1, 0xe4, 0x59, 0x00, 0x03, 0x6d, 0x73,
-  0x67, 0x2e, 0x74, 0x6d, 0x70, 0x6c, 0x00, 0x55, 0x90, 0xbd, 0x6e, 0xc3,
-  0x30, 0x0c, 0x84, 0xe7, 0xe8, 0x29, 0x58, 0x4d, 0xed, 0x60, 0x69, 0xef,
-  0xd6, 0x36, 0x2d, 0xea, 0x25, 0x09, 0x90, 0xbc, 0x00, 0x23, 0xd3, 0x82,
-  0x10, 0xeb, 0x07, 0x94, 0x32, 0x18, 0x46, 0xde, 0xbd, 0x52, 0xdc, 0x1f,
-  0x74, 0x23, 0x8f, 0xdf, 0x1d, 0x0e, 0xd4, 0x1a, 0xde, 0xe2, 0x40, 0x60,
-  0x29, 0x10, 0x63, 0xa1, 0x01, 0xce, 0x33, 0x70, 0xcc, 0x5d, 0x15, 0x3a,
-  0x1b, 0x95, 0xd0, 0x1a, 0x72, 0xbc, 0xb2, 0xa1, 0x67, 0x58, 0x16, 0x50,
-  0x1f, 0x6e, 0xa2, 0x3e, 0x8c, 0x51, 0xf5, 0xa1, 0x8d, 0xaf, 0x98, 0x09,
-  0x6e, 0xb7, 0x46, 0x6d, 0xf7, 0xb0, 0xdb, 0x9f, 0xe0, 0x7d, 0xdb, 0x9f,
-  0x1e, 0x44, 0x42, 0x73, 0x41, 0x4b, 0xff, 0x2d, 0x87, 0x55, 0xdc, 0xa1,
-  0xbf, 0x7b, 0x84, 0xf3, 0x29, 0x72, 0x81, 0x47, 0xb1, 0x59, 0x96, 0x0e,
-  0xdc, 0x08, 0x91, 0x41, 0x1d, 0x13, 0x19, 0xf5, 0x89, 0xf9, 0x38, 0x39,
-  0x43, 0x7f, 0xeb, 0x0b, 0x33, 0xce, 0xcd, 0xb5, 0x91, 0x14, 0x4c, 0x1c,
-  0x5c, 0xb0, 0xfa, 0xec, 0x02, 0xf2, 0x2c, 0xab, 0x34, 0xfa, 0x22, 0xd7,
-  0x14, 0x0a, 0xc3, 0x4a, 0xb9, 0x28, 0x85, 0x80, 0x56, 0x80, 0x31, 0xd8,
-  0x9f, 0xa4, 0xef, 0x0a, 0xb9, 0x31, 0x00, 0xb2, 0xd5, 0xab, 0xa3, 0xbc,
-  0x83, 0xbf, 0xe6, 0x27, 0x21, 0xea, 0xa1, 0x90, 0x4f, 0x53, 0xfd, 0x08,
-  0x48, 0x9f, 0xad, 0x4a, 0xc8, 0xc5, 0xe1, 0xa4, 0x4a, 0x15, 0xe5, 0x1a,
-  0xd6, 0xd0, 0x2f, 0x3a, 0xb2, 0xfc, 0x15, 0x3f, 0x01, 0x00, 0x00 }
-
